fixed a couple compiler warnings on Mac; updated 2.0.1 release date & notes

+++ projects/tophat/trunk/src/common.h
@@ -588,7 +588,7 @@ class GBamWriter {
 	   pre_pos = post_block_addr << 16;
 	 }
 
-       fprintf(findex, "%ld\t%ld\n", read_id, pre_pos);
+       fprintf(findex, "%ld\t%ld\n", (long)read_id, (long)pre_pos);
        idxcount = 0;
      }
    }
